chore: move stripe iframe style helper
